requirements.txt
asgiref
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.3
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
billiard==4.2.0
bleach==6.0.0
blinker==1.7.0
Django
Flask
django-cors-headers
djangorestframework
django-ratelimit
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
